fix: use dropdown for LUIS region in skill bot settings (#5318)
* start making field generic

* Create FieldWithCustomButton.tsx

* refactoring and spelling fix

* move props and use generic field
